requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
atlassian-python-api==3.37.0
attrs==21.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==4.4.2
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
filelock==3.6.0
frozenlist==1.3.3
github==1.2.7
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyte==0.8.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ulauncher==5.15.3
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.15.0
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
